requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
apify-client>=1.6.0
google-generativeai>=0.7.0